ovn: move management port operations to ovn package
